Refactor Label, HighlightedLabel
