+++ crates/ui/src/components.rs
@@ -6,6 +6,7 @@ mod disclosure;
 mod divider;
 mod dropdown_menu;
 mod facepile;
+mod highlighted_label;
 mod icon;
 mod image;
 mod indent_guides;
@@ -42,6 +43,7 @@ pub use disclosure::*;
 pub use divider::*;
 pub use dropdown_menu::*;
 pub use facepile::*;
+pub use highlighted_label::*;
 pub use icon::*;
 pub use image::*;
 pub use indent_guides::*;
